--- Breeze.UI/src/app/wallet/advanced/advanced-ico/advanced-ico.component.html
@@ -1,40 +1,0 @@
-<form [formGroup]='icoFormGroup'>
-    <div class="container" style="border:1px solid lightgray; border-radius: 4px; padding:10px; position: absolute; left:100px; width:65%">
-        <h4>ICO</h4>
-        <div style="margin-top:15px">
-            
-            <div>
-                <label style="font-size:14px;margin-bottom:0px">Extended Public Key</label>
-                <div class="myAddress" *ngIf="extPubKeyLoadingState.success; else elseFeedback"><code style="overflow-wrap: break-word">{{ extPubKey }}</code></div>
-                <ng-template #elseFeedback>
-                    <app-feedback [loading]="extPubKeyLoadingState.loading" 
-                                  [errored]="extPubKeyLoadingState.errored"
-                                  [erroredText]="'Failed to get Extended Public Key'"></app-feedback>
-                </ng-template>
-            </div>
-
-            <div style="margin-top:20px">
-                <label style="font-size:14px;margin-bottom:1px">Generate Addresses</label>
-                <div style="display:flex; flex-direction:row">
-                    <div class="input-group" style="height:35px; margin-bottom:1px; width:240px">
-                        <input formControlName="addressCountControl" type="text" class="form-control" placeholder="Number to generate..." style="border-radius: 0px">
-                        <span class="input-group-btn">
-                            <div>
-                                <button *ngIf="!addressCountControl.invalid && addressCount" (click)="generateAddresses()"
-                                        class="btn btn-default" type="button">Go</button>
-                                <button *ngIf="addressCountControl.invalid || !addressCount" style="color: gray; outline:none" class="btn btn-default" type="button">Go</button>
-                            </div>
-                        </span>
-                    </div>
-                    <img style="width:16px; height:16px; margin-left:6px" 
-                         *ngIf="showTick" src="../../../../assets/images/Tick_Mark-16.png"/>
-                    <app-feedback *ngIf="!generateAddressesLoadingState.success" style="margin-left:5px; width:115px"
-                        [loading]="generateAddressesLoadingState.loading" 
-                        [errored]="generateAddressesLoadingState.errored"
-                        [erroredText]="'Failed to generate'"></app-feedback>
-                </div>
-            </div>
-
-        </div>
-    </div>
-</form>
